feat: integração da biblioteca do chatbot na implementãção via RMI
